backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dtw-python==1.7.2
exceptiongroup==1.3.0
fastapi==0.119.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.217
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
